produtos/urls.py
from django.urls import path

from .views import inicio, listar_locais, adicionar_local

urlpatterns = [
    path('', inicio, name='inicio'),
    path('locais/', listar_locais, name='listar_locais'),
    path('locais/adicionar/', adicionar_local, name='adicionar_local'),

]
